refactor: modularize file and grep distillation into dedicated submodules
