+++ src/hooks/post_tool.rs
@@ -47,13 +47,13 @@ pub fn process_payload(
             } else {
                 &normalized.command
             };
-            return process_file_read(&content, filepath).map(wrap_hook_output);
+            return distill_readfile(&content, filepath).map(wrap_hook_output);
         }
         "Grep" => {
             if !agent_config.grep_enabled() {
                 return None;
             }
-            return process_grep_output(&content).map(wrap_hook_output);
+            return distill_grep(&content).map(wrap_hook_output);
         }
         "WebFetch" => {
             if !agent_config.webfetch_enabled() {
@@ -408,285 +408,8 @@ fn wrap_hook_output(distilled: String) -> String {
 
 // ── NON-BASH TOOL DISTILLATION ───────────────────────────────────────
 
-fn process_file_read(content: &str, filepath: &str) -> Option<String> {
-    let line_count = content.lines().count();
-    if line_count < 50 {
-        return None; // Small files pass through
-    }
-
-    let ext = std::path::Path::new(filepath)
-        .extension()
-        .and_then(|e| e.to_str())
-        .unwrap_or("");
-
-    let distilled = match ext {
-        "rs" => distill_rust_file(content),
-        "py" => distill_python_file(content),
-        "ts" | "tsx" | "js" | "jsx" => distill_js_ts_file(content),
-        "go" => distill_go_file(content),
-        "java" | "kt" => distill_java_file(content),
-        "json" => distill_json_file(content),
-        "toml" | "yaml" | "yml" => distill_config_file(content, ext),
-        "log" | "txt" => distill_log_file(content),
-        _ => distill_unknown_file(content),
-    };
-
-    // Only return if meaningful compression achieved
-    if distilled.len() < content.len() * 8 / 10 {
-        Some(format!(
-            "[OMNI ReadFile: {} → distilled ({} lines)]\n{}",
-            filepath, line_count, distilled
-        ))
-    } else {
-        None
-    }
-}
-
-fn distill_rust_file(content: &str) -> String {
-    let mut out = String::new();
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with("pub fn ")
-            || trimmed.starts_with("pub async fn ")
-            || trimmed.starts_with("pub struct ")
-            || trimmed.starts_with("pub enum ")
-            || trimmed.starts_with("pub trait ")
-            || trimmed.starts_with("impl ")
-            || trimmed.starts_with("pub mod ")
-            || trimmed.starts_with("use ")
-            || trimmed.starts_with("//!")
-            || trimmed.contains("todo!")
-            || trimmed.contains("unimplemented!")
-            || trimmed.contains("panic!")
-        {
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    if out.is_empty() {
-        distill_unknown_file(content)
-    } else {
-        out.trim().to_string()
-    }
-}
-
-fn distill_python_file(content: &str) -> String {
-    let mut out = String::new();
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with("def ")
-            || trimmed.starts_with("async def ")
-            || trimmed.starts_with("class ")
-            || trimmed.starts_with("import ")
-            || trimmed.starts_with("from ")
-            || trimmed.starts_with('@')
-        {
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    if out.is_empty() {
-        distill_unknown_file(content)
-    } else {
-        out.trim().to_string()
-    }
-}
-
-fn distill_js_ts_file(content: &str) -> String {
-    let mut out = String::new();
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with("export ")
-            || trimmed.starts_with("function ")
-            || trimmed.starts_with("class ")
-            || trimmed.starts_with("interface ")
-            || trimmed.starts_with("type ")
-            || (trimmed.starts_with("const ") && trimmed.contains("=>"))
-            || trimmed.starts_with("import ")
-        {
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    if out.is_empty() {
-        distill_unknown_file(content)
-    } else {
-        out.trim().to_string()
-    }
-}
-
-fn distill_go_file(content: &str) -> String {
-    let mut out = String::new();
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with("func ")
-            || trimmed.starts_with("type ")
-            || trimmed.starts_with("var ")
-            || trimmed.starts_with("const ")
-            || trimmed.starts_with("package ")
-            || trimmed.starts_with("import")
-        {
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    if out.is_empty() {
-        distill_unknown_file(content)
-    } else {
-        out.trim().to_string()
-    }
-}
-
-fn distill_java_file(content: &str) -> String {
-    let mut out = String::new();
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if (trimmed.contains("class ")
-            || trimmed.contains("interface ")
-            || trimmed.contains("public ")
-            || trimmed.contains("private ")
-            || trimmed.contains("protected ")
-            || trimmed.starts_with("import ")
-            || trimmed.starts_with("package "))
-            && !trimmed.starts_with("//")
-            && !trimmed.is_empty()
-        {
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    if out.is_empty() {
-        distill_unknown_file(content)
-    } else {
-        out.trim().to_string()
-    }
-}
-
-fn distill_json_file(content: &str) -> String {
-    let lines: Vec<&str> = content.lines().collect();
-    let total = lines.len();
-    if total <= 30 {
-        return content.trim().to_string();
-    }
-    let head: Vec<&str> = lines.iter().take(15).copied().collect();
-    format!(
-        "{}\n... [{} more lines — full JSON in RewindStore]",
-        head.join("\n"),
-        total - 15
-    )
-}
-
-fn distill_config_file(content: &str, ext: &str) -> String {
-    let lines: Vec<&str> = content.lines().collect();
-    let total = lines.len();
-    if total <= 40 {
-        return content.trim().to_string();
-    }
-    let mut out = String::new();
-    for line in &lines {
-        let trimmed = line.trim();
-        if (ext == "toml"
-            && (trimmed.starts_with('[')
-                || (!trimmed.starts_with('#')
-                    && trimmed.contains('=')
-                    && !trimmed.starts_with(' '))))
-            || (matches!(ext, "yaml" | "yml")
-                && !trimmed.starts_with(' ')
-                && !trimmed.starts_with('#')
-                && trimmed.ends_with(':'))
-        {
-            out.push_str(line);
-            out.push('\n');
-        }
-    }
-    if out.is_empty() {
-        distill_unknown_file(content)
-    } else {
-        format!("[Config structure — {} lines total]\n{}", total, out.trim())
-    }
-}
-
-fn distill_log_file(content: &str) -> String {
-    let mut errors = 0usize;
-    let mut warnings = 0usize;
-    let mut error_lines: Vec<String> = vec![];
-    for line in content.lines() {
-        let l = line.to_lowercase();
-        if l.contains("error") || l.contains("fatal") || l.contains("panic") {
-            errors += 1;
-            error_lines.push(line.to_string());
-        } else if l.contains("warn") {
-            warnings += 1;
-        }
-    }
-    let total = content.lines().count();
-    let mut out = format!(
-        "Log: {} errors, {} warnings ({} total lines)\n",
-        errors, warnings, total
-    );
-    for err in error_lines.iter().take(10) {
-        out.push_str(err);
-        out.push('\n');
-    }
-    if errors > 10 {
-        out.push_str(&format!("... [{} more error lines]\n", errors - 10));
-    }
-    out.trim().to_string()
-}
-
-fn distill_unknown_file(content: &str) -> String {
-    let lines: Vec<&str> = content.lines().collect();
-    let total = lines.len();
-    if total <= 30 {
-        return content.trim().to_string();
-    }
-    let head: Vec<&str> = lines.iter().take(15).copied().collect();
-    let tail: Vec<&&str> = lines.iter().rev().take(5).collect();
-    let tail_rev: Vec<&&str> = tail.into_iter().rev().collect();
-    format!(
-        "--- HEAD ({} total lines) ---\n{}\n... [{} lines omitted] ...\n--- TAIL ---\n{}",
-        total,
-        head.join("\n"),
-        total - 20,
-        tail_rev
-            .iter()
-            .map(|l| **l)
-            .collect::<Vec<&str>>()
-            .join("\n")
-    )
-}
-
-fn process_grep_output(content: &str) -> Option<String> {
-    let line_count = content.lines().count();
-    if line_count < 20 {
-        return None;
-    } // Small results pass through
-
-    let files: std::collections::HashSet<&str> = content
-        .lines()
-        .filter_map(|l| l.split(':').next())
-        .filter(|f| !f.is_empty())
-        .collect();
-    let file_count = files.len();
-    let top: Vec<&str> = content.lines().take(15).collect();
-    let summary = format!(
-        "[OMNI Grep: {} matches in {} files]\n{}{}",
-        line_count,
-        file_count,
-        top.join("\n"),
-        if line_count > 15 {
-            format!("\n... [{} more matches]", line_count - 15)
-        } else {
-            String::new()
-        }
-    );
-    if summary.len() < content.len() * 8 / 10 {
-        Some(summary)
-    } else {
-        None
-    }
-}
-
+use crate::distillers::readfile::distill_readfile;
+use crate::distillers::search::distill_grep;
 fn process_web_content(content: &str) -> Option<String> {
     let line_count = content.lines().count();
     if line_count < 30 {
